Clean up EditorSession getting split between modules.

+++ module/render/src/commonTest/kotlin/org/umamo/render/eval/UnkeyedDrawableRenderTest.kt
@@ -1,6 +1,7 @@
 package org.umamo.render.eval
 
 import org.umamo.runtime.model.BlendMode
+import org.umamo.runtime.model.DEFAULT_DRAW_ORDER
 import org.umamo.runtime.model.Drawable
 import org.umamo.runtime.model.DrawableId
 import org.umamo.runtime.model.DrawableMesh
@@ -61,7 +62,7 @@ class UnkeyedDrawableRenderTest {
 		val drawable = Drawable(drawableId, "M", null, BlendMode.Normal, emptyList(), mesh(), geometryGrid = null)
 		val geometry = CpuDeformationEvaluator().evaluate(model(drawable), emptyMap())
 		assertEquals(1f, geometry.opacity[drawableId])
-		assertEquals(DEFAULT_DRAW_ORDER, geometry.drawOrder[drawableId])
+		assertEquals(DEFAULT_DRAW_ORDER.toFloat(), geometry.drawOrder[drawableId])
 	}
 
 	/**
